backend/requirements.txt
fastapi
uvicorn
supabase
pydantic-settings
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
anthropic
instructor
stripe
resend
apscheduler
twilio